Add number of clients option to test runner
Introduces a 'Number of Clients' input field to the UI and updates the test execution logic to support running tests with multiple clients. The run_test and run_test_thread functions now handle the new parameter, passing it to the test runner script. Input validation for the number of clients is also included.